Finish implementation of problem 2: bfs
Created a neighbours attribute for the Actor class, that on
demand finds all its neighbors, meaning all actors that play
in the same movie as them.

Then, breadth first search is implemented, and a somewhat messy
detangling of the actor and movie information is relayed to the
main block, where it is printed out.

Needs some cleaning up, but works fine now.